+++ src/coreclr/src/pal/tests/palsuite/threading/WaitForSingleObject/WFSOExMutexTest/WFSOExMutexTest.cpp
@@ -178,7 +178,7 @@ DWORD PALAPI WaiterProc(LPVOID lpParameter)
     BOOL Alertable;
     DWORD ret;
 
-    Alertable = (BOOL) lpParameter;
+    Alertable = (BOOL)(SIZE_T) lpParameter;
 
     LARGE_INTEGER performanceFrequency;
     if (!QueryPerformanceFrequency(&performanceFrequency))
